begin simplification of codemirror setup
